requirements.txt
flask
filemagic
Markdown
PyGitHub